Don't use bogus event in emailregistration
I was trying to be smart by calling the same registration event as 'register'
when doing EmailRegistration. However, that event is so low-bandwidth that plugins
go fingerpoken in all the attributes and call methods on the passed-in action and
things like that.

So, now we just fall back to using the low-level stuff, catch any
exceptions, and feel happy. Some stuff might not work, but it's
generally anti-spam stuff more suited to public sites.