Option to pre-load a plugin with checkschema script
One of the problems we've had with running large-scale hosting systems
for StatusNet is enabling new plugins. If the plugin is not enabled,
its database tables are not checked at script time. Conversely, if it
is enabled, it may take several hours to run checkschema for tens of
thousands of sites -- during which time users might see DB errors.

A new argument to checkschema lets it pre-load one or more plugins
before checking the schema. This lets us prepare the plugins' database
tables before they're used in production. In a multihome environment,
this can be combined with tags to gradually roll out a new plugin.

In the config file, a stanza like:

   $site = Status_network::getFromHostname(...);

   if ($site->hasTag('fooenabled')) {
      addPlugin('Foo');
   }

...will only enable the plugin on certain sites. Meanwhile, a bash
script like this should gradually enable the plugin:

   # For all sites...
   for site in `php allsites.php`; do
       # Update the schema for the Foo plugin
       php checkschema.php -s$site.wildcard -xFoo;
       # Enable the Foo plugin
       php settag.php -s$site.wildcard fooenabled;
   done